refactor: Drop support for pglite as it is not used and not planned to be used anymore
