+++ .changeset/swift-suits-search.md
@@ -1,0 +1,5 @@
+---
+"@nmi-agro/fdm-core": patch
+---
+
+Drop support for pglite as it is not used and not planned to be used anymore
