+++ fdm-core/src/cultivation.test.ts
@@ -20,7 +20,7 @@ import {
 } from "./cultivation"
 import * as schema from "./db/schema"
 import { addFarm } from "./farm"
-import type { FdmType } from "./fdm"
+import type { FdmType } from "./fdm.types"
 import { createFdmServer } from "./fdm-server"
 import type { FdmServerType } from "./fdm-server.types"
 import {
