+++ fdm-core/src/farm.ts
@@ -8,12 +8,16 @@ import {
     revokePrincipal,
     updateRole,
 } from "./authorization"
-import type { PrincipalId, PrincipalWithRoles, Role } from "./authorization.types"
+import type {
+    PrincipalId,
+    PrincipalWithRoles,
+    Role,
+} from "./authorization.types"
 import * as schema from "./db/schema"
 import * as authNSchema from "./db/schema-authn"
 import * as authZSchema from "./db/schema-authz"
 import { handleError } from "./error"
-import type { FdmType } from "./fdm"
+import type { FdmType } from "./fdm.types"
 import { removeField } from "./field"
 import { createId } from "./id"
 import {
