--- fdm-core/src/fdm-local.types.d.ts
@@ -1,3 +1,0 @@
-import type { drizzle } from "drizzle-orm/pglite"
-
-export type FdmLocalType = ReturnType<typeof drizzle>
