--- fdm-core/src/fdm.d.ts
@@ -1,5 +1,0 @@
-import type { FdmLocalType } from "./fdm-local.types"
-import type { FdmServerType } from "./fdm-server.types"
-
-// Define type of Fdm
-export type FdmType = FdmServerType | FdmLocalType
