+++ fdm-core/src/fdm.types.d.ts
@@ -1,0 +1,4 @@
+import type { FdmServerType } from "./fdm-server.types"
+
+// Define type of Fdm
+export type FdmType = FdmServerType
