+++ fdm-core/src/harvest.ts
@@ -9,7 +9,7 @@ import { checkPermission } from "./authorization"
 import type { PrincipalId } from "./authorization.types"
 import * as schema from "./db/schema"
 import { handleError } from "./error"
-import type { FdmType } from "./fdm"
+import type { FdmType } from "./fdm.types"
 import type {
     Harvest,
     HarvestParameters,
