+++ fdm-core/src/principal.test.ts
@@ -2,7 +2,7 @@ import { eq } from "drizzle-orm"
 import { beforeAll, describe, expect, inject, it } from "vitest"
 import { type BetterAuth, createFdmAuth } from "./authentication"
 import * as authNSchema from "./db/schema-authn"
-import type { FdmType } from "./fdm"
+import type { FdmType } from "./fdm.types"
 import { createFdmServer } from "./fdm-server"
 import { createId } from "./id"
 import { getPrincipal, identifyPrincipal, lookupPrincipal } from "./principal"
