+++ fdm-core/src/principal.ts
@@ -1,7 +1,7 @@
 import { eq, ilike, or } from "drizzle-orm"
 import * as authNSchema from "./db/schema-authn"
 import { handleError } from "./error"
-import type { FdmType } from "./fdm"
+import type { FdmType } from "./fdm.types"
 import type { Principal } from "./principal.types"
 
 /**
